requirements.txt
APScheduler==3.10.4
pytest==8.3.2
python-dotenv==1.0.1
PyYAML==6.0.1
SQLAlchemy==2.0.31